Agent: Move build_monkey_commandline_parameters to command_builders
