+++ monkey/infection_monkey/command_builders/__init__.py
@@ -1,3 +1,4 @@
 from .agent_command_builder_factory import AgentCommandBuilderFactory
 from .linux_agent_command_builder import LinuxAgentCommandBuilder
 from .windows_agent_command_builder import WindowsAgentCommandBuilder
+from .utils import build_monkey_commandline_parameters
