+++ monkey/infection_monkey/command_builders/utils.py
@@ -1,7 +1,8 @@
 from pathlib import PurePath
-from typing import Optional, Union
+from typing import Optional, Sequence, Union
 
 from agentpluginapi import DropperExecutionMode, LinuxRunOptions, WindowsRunOptions
+from monkeytypes import AgentID
 
 from infection_monkey.model import DROPPER_ARG, MONKEY_ARG
 
@@ -25,3 +26,27 @@ def get_agent_location(
             else run_options.agent_destination_path
         )
     return destination_path
+
+
+def build_monkey_commandline_parameters(
+    parent: Optional[AgentID] = None,
+    servers: Optional[Sequence[str]] = None,
+    depth: Optional[int] = None,
+    location: Union[str, PurePath, None] = None,
+) -> Sequence[str]:
+    cmdline = []
+
+    if parent is not None:
+        cmdline.append("-p")
+        cmdline.append(str(parent))
+    if servers:
+        cmdline.append("-s")
+        cmdline.append(",".join(servers))
+    if depth is not None:
+        cmdline.append("-d")
+        cmdline.append(str(depth))
+    if location is not None:
+        cmdline.append("-l")
+        cmdline.append(str(location))
+
+    return cmdline
