+++ monkey/tests/unit_tests/infection_monkey/utils/test_commands.py
@@ -1,6 +1,5 @@
 from ipaddress import IPv4Address
 from typing import Callable
-from uuid import UUID
 
 import pytest
 from agentpluginapi import TargetHost
@@ -10,7 +9,6 @@ from infection_monkey.utils.commands import (
     build_agent_download_command,
     build_dropper_script_download_command,
     build_monkey_commandline,
-    build_monkey_commandline_parameters,
     get_monkey_commandline_linux,
     get_monkey_commandline_windows,
 )
@@ -20,38 +18,6 @@ from infection_monkey.utils.ids import get_agent_id
 @pytest.fixture
 def agent_id():
     return get_agent_id()
-
-
-def test_build_monkey_commandline_parameters_arguments():
-    agent_id = UUID("9614480d-471b-4568-86b5-cb922a34ed8a")
-    expected = [
-        "-p",
-        str(agent_id),
-        "-s",
-        "127.127.127.127:5000,138.138.138.138:5007",
-        "-d",
-        "0",
-        "-l",
-        "C:\\windows\\abc",
-    ]
-    actual = build_monkey_commandline_parameters(
-        agent_id,
-        ["127.127.127.127:5000", "138.138.138.138:5007"],
-        0,
-        "C:\\windows\\abc",
-    )
-
-    assert expected == actual
-
-
-def test_build_monkey_commandline_parameters_depth_condition_greater():
-    expected = [
-        "-d",
-        "50",
-    ]
-    actual = build_monkey_commandline_parameters(depth=50)
-
-    assert expected == actual
 
 
 def test_get_monkey_commandline_windows():
